+++ R/meta-chk.R
@@ -362,7 +362,7 @@ chkglmm <- function(sm, method.tau, method.random.ci, method.predict,
     stop("Hartung-Knapp method for prediction interval not ",
          "available for GLMMs; use 'method.predict = \"V\".",
          call. = FALSE)
-
+  
   return(invisible(NULL))
 }
 
@@ -384,6 +384,44 @@ chkmlm <- function(method.tau, missing.method.tau,
          "three-level models: \"V\", \"HTS\", \"S\".",
          call. = FALSE)
   #
+  return(invisible(NULL))
+}
+
+chklrp <- function(sm, method.tau, method.random.ci, method.predict,
+                   adhoc.hakn.ci, adhoc.hakn.pi,
+                   sm.allowed) {
+  
+  if (!(sm %in% sm.allowed))
+    stop("Penalised logistic regression only possible with ",
+         "argument 'sm = \"", paste0(sm.allowed, collapse = ", "), "\"'.",
+         call. = FALSE)
+  ##
+  if (any(method.random.ci == "KR"))
+    stop("Kenward-Roger method for random effects meta-analysis not ",
+         "available for penalised logistic regression.",
+         call. = FALSE)
+  ##
+  if (any(method.random.ci == "HK" & adhoc.hakn.ci != ""))
+    stop("Hartung-Knapp method with ad hoc correction not ",
+         "available for penalised logistic regression.",
+         call. = FALSE)
+  ##
+  if (any(method.predict == "KR"))
+    stop("Kenward-Roger method for prediction interval not ",
+         "available for penalised logistic regression.",
+         call. = FALSE)
+  ##
+  if (any(method.predict == "NNF"))
+    stop("Bootstrap method for prediction interval not ",
+         "available for penalised logistic regression.",
+         call. = FALSE)
+  ##
+  if (any(method.predict == "HK")) 
+    stop("Hartung-Knapp method for prediction interval not ",
+         "available for penalised logistic regression; use ",
+         "'method.predict = \"V\".",
+         call. = FALSE)
+  
   return(invisible(NULL))
 }